WHAN-33 moved util class to existing package and make it package private. also add some javadoc examples for request attribute values
