+++ media/changes.xml
@@ -23,6 +23,12 @@
     xsi:schemaLocation="http://maven.apache.org/changes/1.0.0 http://maven.apache.org/plugins/maven-changes-plugin/xsd/changes-1.0.0.xsd">
   <body>
 
+    <release version="1.9.0" date="no-released">
+      <action type="update" dev="mrozati" issue="WHAN-33">
+        ResourceMedia supports request attributes for responsive image sizes. This can be either `imageSizes` and `widthOptions` (width + required flag), or `pictureSourceMediaFormat`, `pictureSourceMedia` and `pictureSourceWidths`
+      </action>
+    </release>
+
     <release version="1.8.2" date="2020-01-30">
       <action type="update" dev="sseifert">
         Detect media dimensions for original renditions for non-image binaries (from asset metadata).
@@ -35,9 +41,6 @@
       </action>
       <action type="remove" dev="sseifert">
         DamVideoMediaMarkupBuilder: Do not generate flash player as fallback.
-      </action>
-      <action type="update" dev="mrozati" issue="WHAN-33">
-        ResourceMedia supports request attributes for responsive image sizes. This can be either `imageSizes` and `widthOptions` (width + required flag), or `pictureSourceMediaFormat`, `pictureSourceMedia` and `pictureSourceWidths`
       </action>
     </release>
