+++ media/src/main/java/io/wcm/handler/media/ui/ImageUtils.java
@@ -1,5 +1,23 @@
-/* Copyright (c) pro!vision GmbH. All rights reserved. */
-package io.wcm.handler.media.util;
+/*
+ * #%L
+ * wcm.io
+ * %%
+ * Copyright (C) 2014 wcm.io
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+package io.wcm.handler.media.ui;
 
 import java.util.Arrays;
 
@@ -15,7 +33,7 @@ import io.wcm.handler.media.MediaBuilder;
 import io.wcm.handler.media.format.MediaFormat;
 import io.wcm.handler.media.format.MediaFormatHandler;
 
-public class ImageUtils {
+class ImageUtils {
 
   private static final Logger LOG = LoggerFactory.getLogger(ImageUtils.class);
